+++ app/src/androidTest/java/org/simple/clinic/patient/PatientRepositoryAndroidTest.kt
@@ -72,6 +72,7 @@ import java.util.Optional
 import java.util.UUID
 import javax.inject.Inject
 import javax.inject.Named
+import org.simple.clinic.patient.Answer as PatientAnswer
 
 class PatientRepositoryAndroidTest {
 
@@ -920,7 +921,7 @@ class PatientRepositoryAndroidTest {
       ageDetails = ageDetails,
       patientRecordedAt = this.recordedAt,
       updatedAt = recordedAt,
-      isEligibleForReassignment = isEligibleForReassignment,
+      eligibleForReassignment = eligibleForReassignment,
   )
 
   private fun verifyRecentPatientOrder(
@@ -975,7 +976,7 @@ class PatientRepositoryAndroidTest {
           ageDetails = ageDetails,
           patientRecordedAt = this.recordedAt,
           updatedAt = updatedAt,
-          isEligibleForReassignment = isEligibleForReassignment,
+          eligibleForReassignment = eligibleForReassignment,
       )
     }
   }
@@ -1117,7 +1118,7 @@ class PatientRepositoryAndroidTest {
           ageDetails = ageDetails,
           patientRecordedAt = this.recordedAt,
           updatedAt = createdAt,
-          isEligibleForReassignment = isEligibleForReassignment,
+          eligibleForReassignment = eligibleForReassignment,
       )
     }
   }
@@ -4502,20 +4503,24 @@ class PatientRepositoryAndroidTest {
         patientStatus = Active,
         patientCreatedAt = Instant.parse("2017-01-01T00:00:00Z"),
         patientUpdatedAt = Instant.parse("2018-01-01T00:00:00Z"),
-        patientDeletedAt = null
+        patientDeletedAt = null,
+        eligibleForReassignment = PatientAnswer.Unanswered,
     )
 
     facilityRepository.save(listOf(facility))
     patientRepository.save(listOf(patientProfile))
 
+    val patientEligibleForReassignmentBeforeUpdate = patientRepository.patientImmediate(patientId)?.eligibleForReassignment ?: false
+    assertThat(patientEligibleForReassignmentBeforeUpdate).isEqualTo(PatientAnswer.Unanswered)
+
     // when
     patientRepository.updatePatientReassignmentEligibilityStatus(
         patientUuid = patientId,
-        isEligibleForReassignment = true
+        eligibleForReassignment = PatientAnswer.Yes
     )
 
     // then
-    val isPatientEligibleForReassignment = patientRepository.patientImmediate(patientId)?.isEligibleForReassignment ?: false
-    assertThat(isPatientEligibleForReassignment).isTrue()
+    val patientEligibleForReassignmentAfterUpdate = patientRepository.patientImmediate(patientId)?.eligibleForReassignment ?: false
+    assertThat(patientEligibleForReassignmentAfterUpdate).isEqualTo(PatientAnswer.Yes)
   }
 }