+++ app/src/main/java/org/simple/clinic/patient/onlinelookup/api/LookupPatientOnline.kt
@@ -6,6 +6,7 @@ import org.simple.clinic.drugs.PrescribedDrug
 import org.simple.clinic.medicalhistory.Answer
 import org.simple.clinic.medicalhistory.MedicalHistory
 import org.simple.clinic.overdue.Appointment
+import org.simple.clinic.patient.Answer.*
 import org.simple.clinic.patient.CompleteMedicalRecord
 import org.simple.clinic.patient.Patient
 import org.simple.clinic.patient.PatientAgeDetails
@@ -114,7 +115,7 @@ class LookupPatientOnline @Inject constructor(
             registeredFacilityId = response.registeredFacilityId,
             assignedFacilityId = response.assignedFacilityId,
             retainUntil = retainUntil,
-            isEligibleForReassignment = false,
+            eligibleForReassignment = Unanswered,
         ),
         address = response.address.toDatabaseModel(),
         phoneNumbers = phoneNumbers,